backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==13.0
graphiti-core[anthropic]==0.7
neo4j==5.26.0
anthropic==0.43.0
python-dotenv==1.0.1
pydantic==2.10.0
